requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
alembic==1.13.1
fastapi==0.110.0
h11==0.14.0
idna==3.6
pydantic==2.6.1
pydantic_core==2.16.2
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
httpx==0.27.0
pytest==8.0.1
pytest-asyncio==0.23.5
python-dotenv==1.0.0
flake8==7.0.0
pre-commit==3.6.2